+++ changelog.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+### Added
+
+- More fine-grained functions in `Convex.Wallet.Operator` and `Convex.Query` signing & balancing transactions
+
+### Deleted
+
 ## [0.2.0.0-old-node]
 
 ### Changed